+++ ivp_surface_manager/ivp_gridbuild_array.cxx
@@ -813,7 +813,7 @@ void IVP_GridBuilder_Array::convert_array_to_compact_ledges(const IVP_Template_C
 IVP_Compact_Grid *IVP_GridBuilder_Array::compile_ledges_into_compact_grid(const IVP_Template_Compact_Grid *gp, IVP_U_Vector<IVP_Compact_Ledge> *ledges)
 {
     IVP_Compact_Grid *cg = NULL;
-    int buffer_size = (char *)&cg->offset_compact_ledge_array[0] - (char *)cg;  // size for base compact ledge
+    intp buffer_size = (char *)&cg->offset_compact_ledge_array[0] - (char *)cg;  // size for base compact ledge
 
     buffer_size += ledges->len() * sizeof(int);                                     // add buffersize for ledge index array
     buffer_size += (n_rows - 1) * (n_cols - 1) * sizeof(IVP_Compact_Grid_Element);  // the grid referencing ledge index array
